ACCUMULO-4184: Fixed issues with the HostRegexTableLoadBalancer
- Balancer now watches for configuration changes so the master does not have to be restarted
- Fixed a bug where pools overlapped the same hosts and tablets for those tables were constantly reassigned.
- Added a property to control the number of migrations that can be created by this balancer in one pass
- Fix issue where matching properties were being parsed as table regex properties
