+++ server/base/src/main/java/org/apache/accumulo/server/master/balancer/HostRegexTableLoadBalancer.java
@@ -21,9 +21,11 @@ import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Random;
 import java.util.Set;
 import java.util.SortedMap;
 import java.util.TreeMap;
@@ -31,6 +33,7 @@ import java.util.regex.Pattern;
 
 import org.apache.accumulo.core.client.admin.TableOperations;
 import org.apache.accumulo.core.conf.AccumuloConfiguration;
+import org.apache.accumulo.core.conf.ConfigurationObserver;
 import org.apache.accumulo.core.conf.Property;
 import org.apache.accumulo.core.data.impl.KeyExtent;
 import org.apache.accumulo.core.master.thrift.TabletServerStatus;
@@ -38,6 +41,8 @@ import org.apache.accumulo.core.tabletserver.thrift.TabletStats;
 import org.apache.accumulo.server.conf.ServerConfiguration;
 import org.apache.accumulo.server.master.state.TServerInstance;
 import org.apache.accumulo.server.master.state.TabletMigration;
+import org.apache.commons.lang3.builder.ToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
 import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -56,18 +61,24 @@ import org.slf4j.LoggerFactory;
  * <b>table.custom.balancer.host.regex.pool.check=5m</b><br>
  * Regex matching can be based on either the host name (default) or host ip address. To set this balancer to match the regular expressions to the tablet server
  * IP address, then set the following property:<br>
- * <b>table.custom.balancer.host.regex.is.ip=true</b>
+ * <b>table.custom.balancer.host.regex.is.ip=true</b><br>
+ * It's possible that this balancer may create a lot of migrations. To limit the number of migrations that are created during a balance call, set the following
+ * property (default 250):<br>
+ * <b>table.custom.balancer.host.regex.concurrent.migrations</b>
  *
  */
-public class HostRegexTableLoadBalancer extends TableLoadBalancer {
+public class HostRegexTableLoadBalancer extends TableLoadBalancer implements ConfigurationObserver {
 
   private static final Logger LOG = LoggerFactory.getLogger(HostRegexTableLoadBalancer.class);
   public static final String HOST_BALANCER_PREFIX = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.";
-  public static final String HOST_BALANCER_OOB_CHECK = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.oob.period";
+  public static final String HOST_BALANCER_OOB_CHECK_KEY = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.oob.period";
   private static final String HOST_BALANCER_OOB_DEFAULT = "5m";
   public static final String HOST_BALANCER_POOL_RECHECK_KEY = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.pool.check";
   private static final String HOST_BALANCER_POOL_RECHECK_DEFAULT = "1m";
-  public static final String HOST_BALANCER_REGEX_USING_IPS = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.is.ip";
+  public static final String HOST_BALANCER_REGEX_USING_IPS_KEY = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey() + "balancer.host.regex.is.ip";
+  public static final String HOST_BALANCER_REGEX_MAX_MIGRATIONS_KEY = Property.TABLE_ARBITRARY_PROP_PREFIX.getKey()
+      + "balancer.host.regex.concurrent.migrations";
+  private static final int HOST_BALANCER_REGEX_MAX_MIGRATIONS_DEFAULT = 250;
   protected static final String DEFAULT_POOL = "HostTableLoadBalancer.ALL";
 
   protected long oobCheckMillis = AccumuloConfiguration.getTimeInMillis(HOST_BALANCER_OOB_DEFAULT);
@@ -79,6 +90,7 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
   private volatile long lastPoolRecheck = 0;
   private boolean isIpBasedRegex = false;
   private Map<String,SortedMap<TServerInstance,TabletServerStatus>> pools = new HashMap<String,SortedMap<TServerInstance,TabletServerStatus>>();
+  private int maxTServerMigrations = HOST_BALANCER_REGEX_MAX_MIGRATIONS_DEFAULT;
 
   /**
    * Group the set of current tservers by pool name. Tservers that don't match a regex are put into a default pool. This could be expensive in the terms of the
@@ -171,10 +183,15 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
     poolNameToRegexPattern = new HashMap<>();
     for (Entry<String,String> table : t.tableIdMap().entrySet()) {
       tableIdToTableName.put(table.getValue(), table.getKey());
+      conf.getTableConfiguration(table.getValue()).addObserver(this);
       Map<String,String> customProps = conf.getTableConfiguration(table.getValue()).getAllPropertiesWithPrefix(Property.TABLE_ARBITRARY_PROP_PREFIX);
       if (null != customProps && customProps.size() > 0) {
         for (Entry<String,String> customProp : customProps.entrySet()) {
           if (customProp.getKey().startsWith(HOST_BALANCER_PREFIX)) {
+            if (customProp.getKey().equals(HOST_BALANCER_OOB_CHECK_KEY) || customProp.getKey().equals(HOST_BALANCER_POOL_RECHECK_KEY)
+                || customProp.getKey().equals(HOST_BALANCER_REGEX_USING_IPS_KEY) || customProp.getKey().equals(HOST_BALANCER_REGEX_MAX_MIGRATIONS_KEY)) {
+              continue;
+            }
             String tableName = customProp.getKey().substring(HOST_BALANCER_PREFIX.length());
             String regex = customProp.getValue();
             poolNameToRegexPattern.put(tableName, Pattern.compile(regex));
@@ -182,7 +199,7 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
         }
       }
     }
-    String oobProperty = conf.getConfiguration().get(HOST_BALANCER_OOB_CHECK);
+    String oobProperty = conf.getConfiguration().get(HOST_BALANCER_OOB_CHECK_KEY);
     if (null != oobProperty) {
       oobCheckMillis = AccumuloConfiguration.getTimeInMillis(oobProperty);
     }
@@ -190,10 +207,26 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
     if (null != poolRecheckProperty) {
       poolRecheckMillis = AccumuloConfiguration.getTimeInMillis(poolRecheckProperty);
     }
-    String ipBased = conf.getConfiguration().get(HOST_BALANCER_REGEX_USING_IPS);
+    String ipBased = conf.getConfiguration().get(HOST_BALANCER_REGEX_USING_IPS_KEY);
     if (null != ipBased) {
       isIpBasedRegex = Boolean.parseBoolean(ipBased);
     }
+    String migrations = conf.getConfiguration().get(HOST_BALANCER_REGEX_MAX_MIGRATIONS_KEY);
+    if (null != migrations) {
+      maxTServerMigrations = Integer.parseInt(migrations);
+    }
+    LOG.info("{}", this);
+  }
+
+  @Override
+  public String toString() {
+    ToStringBuilder buf = new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE);
+    buf.append("Pool Recheck Interval", this.poolRecheckMillis);
+    buf.append("Tablet Out Of Bounds Check Interval", this.oobCheckMillis);
+    buf.append("Max Tablet Server Migrations", this.maxTServerMigrations);
+    buf.append("Regular Expressions use IPs", this.isIpBasedRegex);
+    buf.append("Pools", this.poolNameToRegexPattern);
+    return buf.toString();
   }
 
   public Map<String,String> getTableIdToTableName() {
@@ -214,6 +247,10 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
 
   public boolean isIpBasedRegex() {
     return isIpBasedRegex;
+  }
+
+  public int getMaxConcurrentMigrations() {
+    return maxTServerMigrations;
   }
 
   @Override
@@ -270,46 +307,57 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
       try {
         // Check to see if a tablet is assigned outside the bounds of the pool. If so, migrate it.
         for (Entry<TServerInstance,TabletServerStatus> e : current.entrySet()) {
-          for (String assignedPool : getPoolNamesForHost(e.getKey().host())) {
-            for (String table : poolNameToRegexPattern.keySet()) {
-              // pool names are the same as table names, except in the DEFAULT case.
-              if (assignedPool.equals(table)) {
-                // If this tserver is assigned to a regex pool, then we can skip checking tablets for this table on this host.
-                continue;
-              }
-              String tid = t.tableIdMap().get(table);
-              if (null == tid) {
-                LOG.warn("Unable to check for out of bounds tablets for table {}, it may have been deleted or renamed.", table);
-                continue;
-              }
-              try {
-                List<TabletStats> outOfBoundsTablets = getOnlineTabletsForTable(e.getKey(), tid);
-                for (TabletStats ts : outOfBoundsTablets) {
-                  KeyExtent ke = new KeyExtent(ts.getExtent());
-                  if (migrations.contains(ke)) {
-                    LOG.debug("Migration for  out of bounds tablet {} has already been requested", ke);
-                    ;
-                    continue;
-                  }
-                  String poolName = getPoolNameForTable(table);
-                  SortedMap<TServerInstance,TabletServerStatus> currentView = currentGrouped.get(poolName);
-                  if (null != currentView) {
-                    TServerInstance nextTS = currentView.firstKey();
-                    LOG.info("Tablet {} is currently outside the bounds of the regex, migrating from {} to {}", ke, e.getKey(), nextTS);
-                    migrationsOut.add(new TabletMigration(ke, e.getKey(), nextTS));
-                  } else {
-                    LOG.warn("No tablet servers online for pool {}, unable to migrate out of bounds tablets", poolName);
-                  }
+          for (String table : poolNameToRegexPattern.keySet()) {
+            // pool names are the same as table names, except in the DEFAULT case.
+            // If this table is assigned to a pool for this host, then move on.
+            if (getPoolNamesForHost(e.getKey().host()).contains(table)) {
+              continue;
+            }
+            String tid = t.tableIdMap().get(table);
+            if (null == tid) {
+              LOG.warn("Unable to check for out of bounds tablets for table {}, it may have been deleted or renamed.", table);
+              continue;
+            }
+            try {
+              List<TabletStats> outOfBoundsTablets = getOnlineTabletsForTable(e.getKey(), tid);
+              Random random = new Random();
+              for (TabletStats ts : outOfBoundsTablets) {
+                KeyExtent ke = new KeyExtent(ts.getExtent());
+                if (migrations.contains(ke)) {
+                  LOG.debug("Migration for  out of bounds tablet {} has already been requested", ke);
+                  continue;
                 }
-              } catch (TException e1) {
-                LOG.error("Error in OOB check getting tablets for table {} from server {}", tid, e.getKey().host(), e);
+                String poolName = getPoolNameForTable(table);
+                SortedMap<TServerInstance,TabletServerStatus> currentView = currentGrouped.get(poolName);
+                if (null != currentView) {
+                  int skip = random.nextInt(currentView.size());
+                  Iterator<TServerInstance> iter = currentView.keySet().iterator();
+                  for (int i = 0; i < skip; i++) {
+                    iter.next();
+                  }
+                  TServerInstance nextTS = iter.next();
+                  LOG.info("Tablet {} is currently outside the bounds of the regex, migrating from {} to {}", ke, e.getKey(), nextTS);
+                  migrationsOut.add(new TabletMigration(ke, e.getKey(), nextTS));
+                  if (migrationsOut.size() > this.maxTServerMigrations) {
+                    break;
+                  }
+                } else {
+                  LOG.warn("No tablet servers online for pool {}, unable to migrate out of bounds tablets", poolName);
+                }
               }
+            } catch (TException e1) {
+              LOG.error("Error in OOB check getting tablets for table {} from server {}", tid, e.getKey().host(), e);
             }
           }
         }
       } finally {
         this.lastOOBCheck = System.currentTimeMillis();
       }
+    }
+
+    if (migrationsOut.size() > 0) {
+      LOG.warn("Not balancing tables due to moving {} out of bounds tablets", migrationsOut.size());
+      return minBalanceTime;
     }
 
     if (migrations != null && migrations.size() > 0) {
@@ -332,9 +380,25 @@ public class HostRegexTableLoadBalancer extends TableLoadBalancer {
         minBalanceTime = tableBalanceTime;
       }
       migrationsOut.addAll(newMigrations);
+      if (migrationsOut.size() > this.maxTServerMigrations) {
+        break;
+      }
     }
 
     return minBalanceTime;
   }
+
+  @Override
+  public void propertyChanged(String key) {
+    parseConfiguration(this.configuration);
+  }
+
+  @Override
+  public void propertiesChanged() {
+    parseConfiguration(this.configuration);
+  }
+
+  @Override
+  public void sessionExpired() {}
 
 }
